refactor(web): move species registry into species_catalog pkg
registry.py (~1k LOC) + shim species_registry_service; tests patch registry module. Refs #344 phase C.

Made-with: Cursor
